+++ crates/openapi/src/models/profile_response.rs
@@ -14,7 +14,7 @@
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct ProfileResponse {
     #[serde(rename = "code", skip_serializing_if = "Option::is_none")]
-    pub code: Option<f32>,
+    pub code: Option<i32>,
     /// the ID of the payment profile. You need to keep track of this value.
     #[serde(rename = "customer_code", skip_serializing_if = "Option::is_none")]
     pub customer_code: Option<String>,
